relocate rex cmd IT
